+++ R/ui_helpers.R
@@ -101,6 +101,7 @@ enabled_analysis_tabs <- function() {
   c(
     reliability = TRUE,
     frequencies = TRUE,
+    paired = TRUE,
     ttest_anova = TRUE,
     correlation = TRUE,
     regression = TRUE,
@@ -122,6 +123,8 @@ app_ui <- function(version) {
     if (isTRUE(analysis_tabs[["reliability"]])) reliability_tab_panel(),
 
     if (isTRUE(analysis_tabs[["frequencies"]])) frequencies_tab_panel(),
+
+    if (isTRUE(analysis_tabs[["paired"]])) paired_tab_panel(),
 
     if (isTRUE(analysis_tabs[["ttest_anova"]])) ttest_anova_tab_panel(),